moved common test fixtures to separate file
